requirements.txt
flask==2.3.3
flask-socketio==5.3.6
pygame==2.5.2
qrcode[pil]==7.4.2
pillow==10.0.1
python-socketio==5.8.0